refactor: Step3 리뷰반영 - 불필요한 코드 정리 & 상수 적용 & input validation 주체 view로 수정
- 불필요한 코드 수정 및 삭제
- 금액 상수 적용
- 로또 금액 Validation을 inputView에서 하는 것으로 수정

+++ src/main/kotlin/lotto/domain/LottoScore.kt
@@ -30,7 +30,7 @@ class LottoScore {
 
     fun rateOfResult(lottoPrice: LottoPrice, lottoResults: List<LottoResult>): BigDecimal {
         val realLottoPrice = lottoPrice / LottoPurchase.LOTTO_PRICE * LottoPurchase.LOTTO_PRICE
-        return BigDecimal(sumOfResult(lottoResults).toDouble()).divide(BigDecimal(realLottoPrice), 2, RoundingMode.HALF_UP)
+        return BigDecimal(sumOfResult(lottoResults)).divide(BigDecimal(realLottoPrice), 2, RoundingMode.HALF_UP)
     }
 
     private fun sumOfResult(lottoResults: List<LottoResult>): Int = lottoResults.sumOf {
